// src/app/kitchen/cabinet-form/model/kitchen-cabinet-type.ts
export enum KitchenCabinetType {
  BASE_ONE_DOOR = 'BASE_ONE_DOOR',
  BASE_TWO_DOOR = 'BASE_TWO_DOOR',
  BASE_OPEN = 'BASE_OPEN',
  BASE_CARGO = 'BASE_CARGO',
  BASE_WITH_DRAWERS = 'BASE_WITH_DRAWERS',
  BASE_SINK = 'BASE_SINK',
  BASE_COOKTOP = 'BASE_COOKTOP',
  BASE_DISHWASHER = 'BASE_DISHWASHER',
  BASE_DISHWASHER_FREESTANDING = 'BASE_DISHWASHER_FREESTANDING',
  BASE_OVEN = 'BASE_OVEN',
  BASE_OVEN_FREESTANDING = 'BASE_OVEN_FREESTANDING',
  BASE_FRIDGE = 'BASE_FRIDGE',
  BASE_FRIDGE_FREESTANDING = 'BASE_FRIDGE_FREESTANDING',
  TALL_CABINET = 'TALL_CABINET',
  PANTRY_PASSAGE = 'PANTRY_PASSAGE',
  CORNER_CABINET = 'CORNER_CABINET',
  UPPER_ONE_DOOR = 'UPPER_ONE_DOOR',
  UPPER_TWO_DOOR = 'UPPER_TWO_DOOR',
  UPPER_OPEN_SHELF = 'UPPER_OPEN_SHELF',
  UPPER_CASCADE = 'UPPER_CASCADE',
  UPPER_HOOD = 'UPPER_HOOD',
  UPPER_DRAINER = 'UPPER_DRAINER'
}